--- fabric-entity-events-v1/src/client/java/net/fabricmc/fabric/mixin/client/entity/event/MinecraftClientMixin.java
@@ -1,37 +1,0 @@
-/*
- * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package net.fabricmc.fabric.mixin.client.entity.event;
-
-import org.jetbrains.annotations.Nullable;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-
-import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.world.ClientWorld;
-
-import net.fabricmc.fabric.api.entity.event.client.ClientWorldChangeEvents;
-
-@Mixin(MinecraftClient.class)
-public class MinecraftClientMixin {
-	@Inject(method = "setWorld", at = @At("TAIL"))
-	private void afterClientWorldChange(@Nullable ClientWorld world, CallbackInfo ci) {
-		MinecraftClient client = (MinecraftClient) (Object) this;
-		ClientWorldChangeEvents.AFTER_CLIENT_WORLD_CHANGE.invoker().afterWorldChange(client, world);
-	}
-}
