+++ fabric-lifecycle-events-v1/src/testmodClient/java/net/fabricmc/fabric/test/event/lifecycle/client/ClientLifecycleTests.java
@@ -16,10 +16,15 @@
 
 package net.fabricmc.fabric.test.event.lifecycle.client;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientLifecycleEvents;
+import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientWorldEvents;
 
 public final class ClientLifecycleTests implements ClientModInitializer {
+	private static final Logger LOGGER = LoggerFactory.getLogger(ClientLifecycleTests.class);
 	private boolean startCalled;
 	private boolean stopCalled;
 
@@ -43,6 +48,12 @@ public final class ClientLifecycleTests implements ClientModInitializer {
 
 			stopCalled = true;
 			System.out.println("Client has started stopping!");
+		});
+
+		ClientWorldEvents.AFTER_CLIENT_WORLD_CHANGE.register((client, world) -> {
+			if (world != null) {
+				LOGGER.info("Client world changed to {}", world.getRegistryKey().getValue());
+			}
 		});
 	}
 }
